refactor: move block template classes to matching package
