+++ src/main/java/de/julianweinelt/ubm/blocks/api/BlockModWall.java
@@ -1,4 +1,4 @@
-package de.julianweinelt.ubm.blocks;
+package de.julianweinelt.ubm.blocks.api;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockWall;
