+++ src/main/java/de/julianweinelt/ubm/blocks/api/BlockRotated.java
@@ -1,10 +1,7 @@
-package de.julianweinelt.ubm.blocks;
+package de.julianweinelt.ubm.blocks.api;
 
-import de.julianweinelt.ubm.blocks.api.EnumHorizontalFacing;
-import de.julianweinelt.ubm.misc.ModCreativeTabs;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
-import net.minecraft.block.properties.PropertyBool;
 import net.minecraft.block.properties.PropertyEnum;
 import net.minecraft.block.state.BlockStateContainer;
 import net.minecraft.block.state.IBlockState;
